lib: Replace Compat<T>::into_result() with set_error_threshold()
Being able to set the error threshold for the whole build chain is much
less verbose and make it easy to change the threshold for a range of the
build.

Making Compat<T> methods return a Result is also a safeguard to not
forget to call .set_last_call_status() for each of them.

Signed-off-by: Mickaël Salaün <mic@digikod.net>