lib: Check BitFlags<AccessFs> compatibility and convert it if needed
Add a TryCompat trait and implement it for BitFlags<AccessFs>.  This
enables to easily implement new BitFlags<T> with the required automatic
downgrade for compatibility purpose.

It is still possible to forge unsupported access rights, but it will
result either to a runtime error or to an automatic compatible
conversion.

Signed-off-by: Mickaël Salaün <mic@digikod.net>