src: Revamp and simplify the compatibility mechanism
This change has two goals:
* Make the compatibility mechanism simpler for this library users and
  developers.
* Only avoid useless Landlock specific calls but run as much code as
  possible in any case: supported system or not.

Replace ErrorThreshold with SupportLevel, switching from four error
thresholds (NoError, Runtime, Incompatible, PartiallyCompatible) to two
levels (Optional and Required).  The NoError is not possible anymore,
because it could hide implementation issues.  Runtime errors are now
always reported and can be handled with common Rust error API.  The
Incompatible and PartiallyCompatible thresholds are replaced with the
Required support level.  Indeed, it doesn't seems useful to
differentiate these kind of errors.  By default, the support level is
set to Optional: best-effort security approach.

All non-Landlock-runtime errors are now returned as early as possible
even for incompatible systems, and this library now duplicates some
kernel checks to help spot misuses that would otherwise be invisible on
unsupported systems (e.g. build a sandbox with inconsistent
access-rights).  There is then no need to store synthetic errors with
the complex CompatObject<T>, CompatBuild<T>, LastCall and Compatibility
types.  They are replaced with a simplified Compatibility type which is
embedded in RulesetInit and RulesetCreated instead of wrapping them in
Compat<T>.  The new CompatState, part of Compatibility, keeps track of
the compatibility state for each ruleset build step.  Removing Compat<T>
has also the benefit to make a much clearer Rust documentation.  As a
downside, it is not possible to change the support level inside the
build chain.  This could be easily implemented with a trait for
RulesetInit and RulesetCreated though.

The PathBeneath type now doesn't directly check compatibility issues on
its own, but instead implements the TryCompat trait which is used by
RulesetCreated to check that all added rules are compatible with the
running system.  This also helps to simplify error handling by removing
it from one layer of the ruleset builder.

Extend ABI enum with a new Unsupported field.  This makes it possible to
easily implement new types (e.g. BitFlags<T>) handling unsupported
systems in a best-effort way (i.e. optional support level).

Add tests for basic compatibility issues.

Explain the rational in the readme.

Signed-off-by: Mickaël Salaün <mic@digikod.net>