ruleset: Call prctl(PR_SET_NO_NEW_PRIVS) when it should work
Call prctl_set_no_new_privs() as long as the kernel knows how to handle
PR_GET_NO_NEW_PRIVS, which appears in Linux 3.5 (July 2012).  Take care
to get a deterministic behavior whether Landlock is supported or not
(e.g. when it is between 3.5 and 5.13, or just when Landlock is not
enabled).

Signed-off-by: Mickaël Salaün <mic@digikod.net>