ruleset: Replace RestrictionStatus with a non-exhaustive struct
Embed the ruleset enforcement status along with the no_new_privs
enforcement status in an non-exhaustive RestrictionStatus struct.  This
enables to easily check for the ruleset and the PR_SET_NO_NEW_PRIVS
enforcement while still being able to extend this returned status with
more details without breaking the crate compatibility.

Add a check in the sandboxer example to error out if Landlock is not
supported by the running kernel.

Signed-off-by: Mickaël Salaün <mic@digikod.net>