compat: Remove useless state update when an error occurs
There is no need to update the compatibility state in case of an error
because this (owned) state would disappear, eaten by the called
function.

It does not make sense to check such state when an error occure in tests
because it will not be visible to the crate users.

Signed-off-by: Mickaël Salaün <mic@digikod.net>