errors: Make all builder methods return the same RulesetError
Remove argument type from RulesetError, which is required to get the
same returned type for every step of a ruleset build.

This is possible because the iterator from RulesetCreated::add_rules()
now returns Result<Self, E> with E implementing From<RulesetError>.
This enables to avoid encapsulating an external error type in
RulesetError and make the use of a ruleset builder call chain much more
easy to handle.  The sandboxer example showcases this behavior with
the new PathEnvError which wraps RulesetError.

Replace two indirect From implementations with explicits map_err() to
avoid stack overflows in tests.

Signed-off-by: Mickaël Salaün <mic@digikod.net>