compat: Add set_compatibility()
The set_compatibility() build method enables to set the compatibility
level of an object builder (e.g. Ruleset).

- set_compatibility(CompatLevel::BestEffort) is like
  set_best_effort(true).

- set_compatibility(CompatLevel::HardRequirement) is like
  set_best_effort(false).

- set_compatibility(CompatLevel::SoftRequirement) makes the ruleset moot
  if one of the following build requests (part of a build method) are
  not supported, but don't return a compatibility error.

CompatLevel::SoftRequirement is particularly useful for access rights
such as AccessFs::Refer.  Indeed, it can be used to ignore the whole
sandboxing if this could break the normal workflow of an application
(which legitimately requires to move files to arbitrary directories).

set_best_effort() is kept for API compatibility.

Signed-off-by: Mickaël Salaün <mic@digikod.net>