compat: Implement Compatible for references too
This enables to change the compatibility for builder objects or their
references like it is done for RulesetAttr and RulesetCreatedAttr.
Make Compatibility a requirement for RulesetAttr and RulesetCreatedAttr
traits to ensure consistency.

This change makes it possible to read the CompatLevel tied to an
object.  It may be superfluous but should not harm.

Signed-off-by: Mickaël Salaün <mic@digikod.net>