tests: Handle kernels with Landlock disabled
It might be useful to run tests with a kernel supporting Landlock but
explicitly disabling it (see the LSM stacking configuration).  In this
case, the kernel returns EOPNOTSUPP which only matches the unstable
ErrorKind::Uncategorized.

Even if tests helpers where only available when building tests, make
them explicitly private outside this crate.

Signed-off-by: Mickaël Salaün <mic@digikod.net>