fs,lib: Add AccessFs::from_file() and test permanent incompatibility
Extend check_ruleset_support() to be able to handle
never-fully-supported rulesets.  This is useful to test permanent
incompatibilities such as directory-only access rights requested on
files, which will be fixed and tested with
too_much_access_rights_for_a_file() in the next commit.

Add AccessFs::From_file(), which indirectly exposes ACCESS_FILE
according to an ABI.  This is useful for this test and other similar
code that don't rely on path_beneath_rules().

Signed-off-by: Mickaël Salaün <mic@digikod.net>