src: Execute each restrict_self() test in a dedicated thread
Ensures restrict_self() is called on a dedicated thread to avoid
inconsistent tests.

The visible effect (with Rust 1.63) was for restrict_self() to sometime
return E2BIG because of too many stacked domains. This inconsistency was
related to the number of tests (i.e. the number of successful
restrict_self() calls), but not directly related to their content.

Signed-off-by: Mickaël Salaün <mic@digikod.net>