ci: Update Ubunty 22.04.3 kernel
GitHub CI updated the Ubuntu kernel with Linux 6.2:
https://ubuntu.com/about/release-cycle#ubuntu-kernel-release-cycle

Reflect this Landlock ABI bump in the tests.

We could only rely on the new landlock-test-tools UML kernels, but this
kind of Ubuntu update should not be frequent, and I'd still like to test
with an Ubuntu kernel. Let's keep it for now, we'll update this CI
configuration from time to time.

Signed-off-by: Mickaël Salaün <mic@digikod.net>