access: Move from_read() and from_write() into AccessFs
The from_read() and from_write() methods make sense for AccessFs but not
for the upcoming AccessNet.

This is a breaking change because the Access trait's signature is
modified.  In practice, it should not be an issue for most users.  Only
those implementing helpers explicitly using Access::from_read() or
Access::from_write() should be concerned.  In this case, they should
just use AccessFs instead.

Fix a typo in the from_all() comment.

Signed-off-by: Mickaël Salaün <mic@digikod.net>