compat,sandboxer: Print hints about Landlock ABI version
Add the LandlockStatus and CurrentABI:
- LandlockStatus can be used to query the running kernel and display
  information about the available Landlock features.
- CurrentABI can be used to compare to known ABI versions, but it cannot
  directly (and should not) be used for other purpose.

Export Debug, Eq, and Ord implementations for ABI, and implement Display
for ABI and CurrentABI.

Replace ABI in Compatibility with LandlockStatus to be able to return
this information with RestrictionStatus.

Update the sandboxer to print hints about the status of Landlock like
for the C example.

Replace some ABI::new_current() calls with LandlockStatus::current().

Signed-off-by: Mickaël Salaün <mic@digikod.net>