access: Document HandledAccess
With commit 554217dda0b7 ("access: Make Access independant from
HandledAccess"), the documentation about RulesetCreated::handle_access()
lists HandledAccess as an argument's requirement, but because this trait
is private, there is no list of types implementing this trait, which is
confusing.

Move private methods to a new PrivateHandledAccess trait and use
HandledAccess as a way to document the argument's requirement.

Fix AddRuleError requirement.

Signed-off-by: Mickaël Salaün <mic@digikod.net>