fs: Add more background to path_beneath_rules() doc
This example is shown at the homepage of rust-landlock's documentation,
and essentially serves as some sort of an introduction to implementers.
Since this example serves as some sort of a "first impression", it is
quite easy for the reader to make a false generalization.

To fix this, some additional background regarding the compatibility
of rules was provided. Additionally, I attempted to separate the role
of file descriptors and paths, as well as add a subtle separation between
Landlock (the LSM) and rust-landlock (the crate).

Signed-off-by: Panagiotis "Ivory" Vasilopoulos <git@n0toose.net>